Added recipeCard.css and recipeCard.js to profile.html and home.html. Created home.js and chefProfile.html
Added chefProfile.html, similar to profile.html but for the chef, added button follow chef for the registered user and sections profile and info. Added home.js shows recipe cards dinamically and added functionality for category buttons (and function to scroll with arrow buttons), if user is registered and has chefs in their followed list shows them in the followed tab.